requirements.txt
lightrag-hku
fastapi
python-dotenv
uvicorn
transformers
tqdm